restructure into single tag with formating options
